// Qanat.Web/src/app/shared/generated/model/custom-attribute-simple-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CustomAttributeSimpleDto { 
    CustomAttributeID?: number;
    GeographyID?: number;
    CustomAttributeTypeID?: number;
    CustomAttributeName?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CustomAttributeSimpleDtoForm { 
    CustomAttributeID?: FormControl<number>;
    GeographyID?: FormControl<number>;
    CustomAttributeTypeID?: FormControl<number>;
    CustomAttributeName?: FormControl<string>;
}

export class CustomAttributeSimpleDtoFormControls { 
    public static CustomAttributeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeographyID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}