// Qanat.Web/src/app/shared/generated/model/custom-attribute-type-simple-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CustomAttributeTypeSimpleDto { 
    CustomAttributeTypeID?: number;
    CustomAttributeTypeName?: string;
    CustomAttributeTypeDisplayName?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CustomAttributeTypeSimpleDtoForm { 
    CustomAttributeTypeID?: FormControl<number>;
    CustomAttributeTypeName?: FormControl<string>;
    CustomAttributeTypeDisplayName?: FormControl<string>;
}

export class CustomAttributeTypeSimpleDtoFormControls { 
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}